fix: Fixed wheel health during TyreSync event
SetVehicleWheelHealth takes a float between 0 and 1000.

Normally the wheel health does not matter too much in the base game so not many people have experienced any issues with it. My (and probably other) realistic wheel damage scripts use it to break off wheels etc.
This would cause issues with it as the wheel health would basically be set to `0` because an integer was used instead of a float.